+++ src/main/java/qna/repository/DeleteHistoryRepository.java
@@ -1,6 +1,7 @@
-package qna.domain;
+package qna.repository;
 
 import org.springframework.data.repository.CrudRepository;
+import qna.domain.DeleteHistory;
 
 public interface DeleteHistoryRepository extends CrudRepository<DeleteHistory, Long> {
