+++ src/main/java/qna/service/DeleteHistoryService.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import qna.domain.DeleteHistory;
-import qna.domain.DeleteHistoryRepository;
+import qna.repository.DeleteHistoryRepository;
 
 import javax.annotation.Resource;
 import java.util.List;
